orders core; inits

+++ Evo 0.1/Evo 0.1/Local Map/MNode.cs
@@ -34,6 +34,6 @@ namespace Mork.Local_Map
 
         public bool explored = true;
 
-        public byte health = 10;
+        public float health = 10;
     }
 }
